Add TestSupport module for test specific deps
This removes the use of Utilities.swift symlinks and adds a tests
specific library which all tests can depend on
