+++ Tests/BuildTests/PackageVersionDataTests.swift
@@ -10,6 +10,7 @@
 
 import XCTest
 
+import TestSupport
 import Basic
 import PackageModel
 import PackageDescription
