+++ Tests/CommandsTests/BuildToolTests.swift
@@ -10,6 +10,7 @@
 
 import XCTest
 
+import TestSupport
 import Basic
 import Commands
