--- Tests/CommandsTests/Utilities.swift
@@ -1,1 +1,0 @@
-../FunctionalTests/Utilities.swift
