+++ Tests/FunctionalTests/SwiftPMXCTestHelperTests.swift
@@ -9,6 +9,7 @@
 */
 
 import Basic
+import TestSupport
 import XCTest
 import Utility
