+++ Tests/FunctionalTests/ValidLayoutTests.swift
@@ -10,6 +10,7 @@
 
 import XCTest
 
+import TestSupport
 import Basic
 import Utility
