+++ Tests/SourceControlTests/CheckoutManagerTests.swift
@@ -9,6 +9,7 @@
 */
 
 import XCTest
+import TestSupport
 import Basic
 import SourceControl
